pyproject.toml
[project]
name = "onionoo-fastapi"
version = "0.1.0"
description = "FastAPI-based semantic/OpenAPI proxy for Tor Onionoo."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  "fastapi[standard]>=0.110",
  "httpx>=0.27",
  "pydantic-settings>=2.2",
  "ruff>=0.6",
  "uvicorn[standard]>=0.27",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

